Issue#300 Spurious Maintenance Message
See description in Issue#300

The problem was that when transitionToNextScreen gets called not all of the product templates have loaded.  When you filter the product groups with groupsWithFilters it can return 0 groups.  This then results in the “Our store is currently undergoing maintenance so no products are available for purchase at this time. Please try again a little later."

This can easily be fixed by checking for [OLProductTemplate isSyncInProgress].  If we are still loading we continue to the products page where we can happily wait for the rest of the products to load.

Perhaps if at the end of loading, no products have been found you can display the maintenance message.